+++ src/graph/graph.h
@@ -33,12 +33,12 @@ float *create_graph(int graph_size, int edge_percentage);
 float *create_locked_graph(int graph_size, int edge_percentage);
 void add_edge(float *graph, int u, int v, float weight);
 void add_null_edge(float *graph, int u, int v);
-unsigned int get_matrix_size(int graph_size);
+unsigned long long get_matrix_size(int graph_size);
 
 float get_edge(float *graph, int u, int v);
 int get_edge_count(float *graph, int graph_size);
 double get_edge_percentage(float *graph, int graph_size);
-unsigned int get_index(int col, int row);
+unsigned long long get_index(int col, int row);
 
 int pick_graph_root(int graph_size);
